requirements/base.txt
gevent>=1.0.1                      
certifi
blinker
pytz
python-dateutil
arrow
python-decouple
click
python-slugify
redis

Flask-Bootstrap
flask-compress
Flask-Admin

#fork for py3 bug - urlparse
https://github.com/gnunicorn/flask-oauth/tarball/master

https://github.com/wooyek/flask-social-blueprint/tarball/develop
Flask-Login>=0.3.0

#Flask-Security
https://github.com/mattupstate/flask-security/tarball/develop

Flask-OAuth
oauth2client
#python-twitter
#facebook-sdk
#google-api-python-client
requests-oauthlib

Flask-Cache
Flask-Moment
Flask-Cache
flask-babelex
flask-babel
Flask-Mail
Flask-Script
Flask-Assets>=0.12
jsmin
cssmin
pycountry
